#13 Fix merge conflict changes
